pyproject.toml
[tool.poetry]
name = "vidutil"
version = "0.0.7"
description = ""
authors = ["Ruben Cronie <rubencronie@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.24.2"
opencv-contrib-python = "4.5.5.64"
pygobject = "3.44.1"
ffmpeg-python = "^0.2.0"
psutil = "^5.9.4"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
autoflake = "^2.0.2"
flake8 = "^6.0.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
